gpu(metal): waitUntilSignaledValue:timeoutMS: takes two uint64_t, not a uint64_t and a uint32_t
The SDK declares the second parameter as uint64_t and the prototype declared it
uint32_t. Arguments are placed by the caller according to the CALLEE's declared
types, so a mismatched width in one of these overloads is the classic way
hand-rolled Objective-C interop corrupts memory, and the whole reason this file
carries a typed overload per selector instead of one variadic declaration.

MsgSendBoolULongUInt becomes MsgSendBoolULongULong, IMetalSharedEvent.WaitUntil
and MetalTimeline.DrainSliceMs carry the corrected width, and the comment that
stated the old signature says what the SDK says. Row 4 absorbs this seam as
declared, so the width has to be right before it does.

+++ KhaozEngine.Gpu.Metal/Internal/IMetalSharedEvent.cs
@@ -51,9 +51,10 @@ namespace KhaozEngine.Gpu.Metal.Internal
         /// </para>
         /// </summary>
         /// <param name="value">The timeline value to wait for.</param>
-        /// <param name="timeoutMs">How long to block before giving up on this attempt.</param>
+        /// <param name="timeoutMs">How long to block before giving up on this attempt. A <c>ulong</c> because
+        /// the SDK's second parameter is a <c>uint64_t</c>, and row 4 absorbs this seam as declared.</param>
         /// <returns>True when the counter reached the value. False when the wait timed out.</returns>
-        bool WaitUntil(ulong value, uint timeoutMs);
+        bool WaitUntil(ulong value, ulong timeoutMs);
 
         /// <summary>
         /// Encode a signal of <paramref name="value"/> on this event into
